feat: implement ArticleBodyReader and update related methods for Yenc header support
