+++ connection.go
@@ -1,8 +1,7 @@
-//go:generate .tools/mockgen -source=./connection.go -destination=./connection_mock.go -package=nntpcli Connection
+//go:generate go tool mockgen -source=./connection.go -destination=./connection_mock.go -package=nntpcli Connection
 package nntpcli
 
 import (
-	"bytes"
 	"errors"
 	"fmt"
 	"io"
@@ -20,101 +19,12 @@ type Connection interface {
 	Authenticate(username, password string) (err error)
 	JoinGroup(name string) error
 	BodyDecoded(msgID string, w io.Writer, discard int64) (int64, error)
-	BodyReader(msgID string) (io.ReadCloser, error)
+	BodyReader(msgID string) (ArticleBodyReader, error)
 	Post(r io.Reader) error
 	CurrentJoinedGroup() string
 	MaxAgeTime() time.Time
 	Stat(msgID string) (int, error)
 	Capabilities() ([]string, error)
-}
-
-type YencHeaders struct {
-	FileName   string
-	FileSize   int64
-	PartNumber int64
-	TotalParts int64
-	Offset     int64
-	PartSize   int64
-	Hash       uint32
-}
-
-type ArticleBodyReader struct {
-	decoder     *rapidyenc.Decoder
-	conn        *connection
-	responseID  uint
-	buffer      *bytes.Buffer
-	headersRead bool
-	yencHeaders *YencHeaders
-	closed      bool
-}
-
-func (r *ArticleBodyReader) Read(p []byte) (n int, err error) {
-	if r.closed {
-		return 0, io.EOF
-	}
-
-	if r.buffer != nil && r.buffer.Len() > 0 {
-		n, err = r.buffer.Read(p)
-		if r.buffer.Len() == 0 {
-			r.buffer = nil
-		}
-		if n > 0 || err != nil {
-			return n, err
-		}
-	}
-
-	return r.decoder.Read(p)
-}
-
-func (r *ArticleBodyReader) GetYencHeaders() (YencHeaders, error) {
-	if r.yencHeaders != nil {
-		return *r.yencHeaders, nil
-	}
-
-	if !r.headersRead {
-		buf := make([]byte, 4096)
-		n, err := r.decoder.Read(buf)
-		if n > 0 {
-			r.buffer = bytes.NewBuffer(buf[:n])
-		}
-		r.headersRead = true
-
-		if err != nil && err != io.EOF {
-			return YencHeaders{}, err
-		}
-	}
-
-	r.yencHeaders = &YencHeaders{
-		FileName:   r.decoder.Meta.FileName,
-		FileSize:   r.decoder.Meta.FileSize,
-		PartNumber: r.decoder.Meta.PartNumber,
-		TotalParts: r.decoder.Meta.TotalParts,
-		Offset:     r.decoder.Meta.Offset,
-		PartSize:   r.decoder.Meta.PartSize,
-		Hash:       r.decoder.Meta.Hash,
-	}
-
-	return *r.yencHeaders, nil
-}
-
-func (r *ArticleBodyReader) Close() error {
-	if r.closed {
-		return nil
-	}
-
-	r.closed = true
-
-	if r.decoder != nil {
-		_, _ = io.Copy(io.Discard, r.decoder)
-		rapidyenc.ReleaseDecoder(r.decoder)
-		r.decoder = nil
-	}
-
-	if r.conn != nil {
-		r.conn.conn.EndResponse(r.responseID)
-	}
-
-	return nil
 }
 
 type connection struct {
@@ -267,7 +177,7 @@ func (c *connection) BodyDecoded(msgID string, w io.Writer, discard int64) (int6
 	return n, nil
 }
 
-func (c *connection) BodyReader(msgID string) (io.ReadCloser, error) {
+func (c *connection) BodyReader(msgID string) (ArticleBodyReader, error) {
 	id, err := c.conn.Cmd("BODY <%s>", msgID)
 	if err != nil {
 		return nil, formatError(err)
@@ -283,7 +193,7 @@ func (c *connection) BodyReader(msgID string) (io.ReadCloser, error) {
 
 	dec := rapidyenc.AcquireDecoder(c.conn.R)
 
-	return &ArticleBodyReader{
+	return &articleBodyReader{
 		decoder:    dec,
 		conn:       c,
 		responseID: id,
